WordPressBot/Blog_Post_1920-1080/requirement.txt
openpyxl
pillow
selenium
pyautogui
gdown